+++ sync-for-commerce/CodatSyncCommerce/Models/Operations/UpdateConnectionAuthorizationResponse.cs
@@ -28,6 +28,11 @@ namespace CodatSyncCommerce.Models.Operations
         public string? ContentType { get; set; } = default!;
 
         /// <summary>
+        /// Your API request was not properly authorized.
+        /// </summary>
+        public ErrorMessage? ErrorMessage { get; set; }
+
+        /// <summary>
         /// HTTP response status code for this operation
         /// </summary>
         public int StatusCode { get; set; } = default!;